Change template

+++ administrator/components/com_j2store/controllers/options.php
@@ -13,18 +13,35 @@ class J2StoreControllerOptions extends F0FController
 
 		$app = JFactory::getApplication();
 		$option_value_id = $app->input->getInt('optionvalue_id');
-		$optionValue = F0FTable::getInstance('Optionvalue','J2StoreTable')->getClone();
-		$optionValue->load($option_value_id);
 
+		$product_optionValue = F0FTable::getInstance('Productoptionvalue','J2StoreTable')->getClone();
+		$product_optionValue->load(array(
+			'optionvalue_id' => $option_value_id
+		));
+		
 		$json  =array();
-		$msg_type = "success";
-		$msg = JText::_('J2STORE_OPTION_VALUE_DELETED_SUCCESSFULLY');
-		$msg_header ='Message';
-		$json['success'] = true;
-		if(!$optionValue->delete()){
+
+		$delete_status = true;
+		if($product_optionValue->j2store_product_optionvalue_id > 0){
+			$delete_status = false;
+		}
+		if($delete_status){
+			$optionValue = F0FTable::getInstance('Optionvalue','J2StoreTable')->getClone();
+			$optionValue->load($option_value_id);
+			$msg_type = "success";
+			$msg_header ='Message';
+			$msg = JText::_('J2STORE_OPTION_VALUE_DELETED_SUCCESSFULLY');
+			$json['success'] = true;
+			if(!$optionValue->delete()){
+				$json['success'] = false;
+				$msg_type = "warning";
+				$msg = JText::_('J2STORE_OPTION_VALUE_DELETE_ERROR');
+				$msg_header ='Warning';
+			}
+		}else{
 			$json['success'] = false;
 			$msg_type = "warning";
-			$msg = JText::_('J2STORE_OPTION_VALUE_DELETE_ERROR');
+			$msg = JText::_('J2STORE_OPTION_VALUE_USED_IN_SOME_PRODUCT');
 			$msg_header ='Warning';
 		}
 
@@ -39,7 +56,7 @@ class J2StoreControllerOptions extends F0FController
 
 	public function getOptions() {
 		$app = JFactory::getApplication();
-		$q = $app->input->post->get('q');
+		$q = $app->input->post->get('q', '','string');
 		$json = array();
 		$model = $this->getThisModel('options');
 		$result = $model->getOptions($q);
